Added basic choose file method for #9:
Export function adds the date/time (do we really need a custom name?)

Import has a dialog that lists all the files in /sdcard/HW-Manager and clicking it imports it. 

I've removed the file extension using Apache IO's FilenameUtils (a trimmed version using only what we need), and the NOTICE file is the Apache 2.0 license for that file.